+++ project/src/constants/colors.ts
@@ -1,0 +1,7 @@
+export default {
+    primary: '#8DA7BE',
+    secodary: '#545F66',
+    tertiary: '#87919E',
+    light: '#DAECF7',
+    dark: '#2D2A32',
+};