+++ project/src/views/BoardView.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import MainTemplate from '../templates/MainTemplate';
+
+const BoardView = () => {
+    return <MainTemplate></MainTemplate>;
+};
+
+export default BoardView;