+++ project/src/views/MainView.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import MainTemplate from '../templates/MainTemplate';
+
+const MainView = () => {
+    return (
+        <MainTemplate>
+            <p>Hi this is main View</p>
+        </MainTemplate>
+    );
+};
+
+export default MainView;