+++ project/Changelog.md
@@ -2,8 +2,10 @@
 
 **Added**
 
+-   Compononents: MainView, MainTemplate, UnderlineBtn, Sidebar
+-   Basics colors of website (in colors.ts);
 -   React-router-dom
--   Basic structure of project
+-   Basic structure of the project
 
 ## V.0.0.1 April 1, 2020